requirements.txt
websocket-client
requests
pytz
python-dotenv
